requirements.txt
flask==3.0.3
flask-socketio==5.3.6
paramiko==3.4.0
python-socketio==5.11.3
requests==2.32.3
pip~=24.1.2
cryptography~=42.0.8
Jinja2~=3.1.4
websocket-client==1.8.0
gevent==24.2.1
gevent-websocket==0.9.5